[nrf fromlist] secure_fw: partitions: Make ITS a weak dependency for ...
... crypto

Internal Trusted Storage is listed as a dependency for the crypto
partition. However if only the RNG or HASH modules are needed from
the crypto partition then ITS is not required.
Excluding ITS partition provides a smaller TF-M image.

Use-case for this is to fit TF-M within 32K on the nrf9160 DK in cases
where only RNG is needed from crypto.

Upstream PR:
https://review.trustedfirmware.org/c/TF-M/trusted-firmware-m/+/13772

Change-Id: Ibb471111f2fbdb3414cff2c7e05331cdcdbc07e1
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>