[nrf toup] platform: nordic_nrf: Add option to initialize hardware to reset on boot
Add an option to the nordic platforms to initialize the hardware state
to the its reset values when the TF-M firmware is booting.
The option is useful when the TF-M firmware image is chain-loaded, for
example, a bootloader, and we need to guarantee that the internal
states of the architecture core blocks are restored to the reset values.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit 6be90b812d6f1f4249237fe18f065acc8738a06f)
(cherry picked from commit 04d7f9ca00b12f811dae80281aa8ae8f76692dd6)