[nrf noup] Fix exception info dump in platform
Cherry-picking the commit "SPM: Move exception info dump to platform"
does not properly include the fault handlers in the final image.
Move the fault.c file from platform_s to tfm_s library as a temporary
workaround.
This is likely because of a missing commit for the build system that is
needed. This commit can be removed when the cherry-picked commit is
included through an upmerge instead.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>