[nrf fromtree] platform: nordic_nrf: Exclude nonsecure storage partition when undefined
Exclude the nonsecure storage partition in nRF platforms when the
storage partition was not defined.
This allows the region to be excluded at build time when the size of the
storage partition is zero.
This simplifies the configuration of the partition in out-of-tree
configuration systems.

Change-Id: I9748293276b0f2cb2b31d7b9519f22a0560e8502
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit 9c44bc931551f027a2666aa680951359890b1756)