[nrf fromtree] platform: nordic_nrf: Add missing ifdef NRF_NS_STORAGE checks
Add missing ifdef NRF_NS_STORAGE check in order to exclude the
non-secure storage partition when this is disabled.
This would otherwise lead to a compilation issues because the start
address is defined but the feature is not enabled, leading to an
inconsistent feature enabled and included check.

Change-Id: I911cd156d9a51bdb75af843177a288d5a3c68b22
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit 5d643d552129ca0f066c96ec67ebf1d7bc79a557)