[nrf noup]: Allow 0 output args for aead_encrypt
This fixes an issue which makes the PSA arch tests
to fail when IPC is enabled. The function
tfm_crypto_aead_encrypt required the outvec
to have at least one output argument. The
PSA arch tests have some invalid cases
where the output buffer is NULL.

See the following line in the PSA arch repository:
api-tests/dev_apis/crypto/test_c024/test_data.h#L186

This changes this behavior to allow zero
parameters for the output which essentially
allows the crypto partition to handle the
error.

This is a noup since in upstream TF-M master
the check is already removed  and the code is
refactored. See commit: 202425a4

Ref: NCSDK-16487

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>