[nrf noup] Workaround to add missing debug info for veneer function in tfm_s image
The linker removes the debug sections of the object files, that
contain functions, which are invoked indirectly via the generated
veneer functions, while the .text section of the function is kept,
 which is correct as it is might be used. This behavior could be
the result of a linker bug, causing ld to misinterpret parts of
the call stack of the veneer functions and or the connected symbols.
As it was not possible to fully identify the root cause a
workaround is introduced. It works by instructing the linker
to include the functions that are called via a veneer function.
This is done by adding the linker option
-Wl,--undefined=function_name for all veneer functions to
the tfm_s target.

As all the xxx_veneer and xxx functions were getting already
included in the hex file before, this workaround only affects
the tfm_s.axf and tfm_s.elf file, which includes the debug
information. Causing the hex files to be the exact same.

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>
(cherry picked from commit 7ec3c913f72409e6e4dcb16bad5edc53436b9a43)