[nrf noup] secure_fw: Add option to log output on a shared UART instance.
Add an option to send the log output from the secure firmware on a
UART instance that would be shared with the non-secure application.

This option is added where the number of UART instances is limited
and the application only cares about the receiving the TF-M log
on fatal errors.

To allow this option to be enabled the log is disabled in the boot
process before the non-secure application is started.
It is enabled again when an unrecoverable exception has occurred in
the secure firmware.

This was not accepted upstream, they have a different way of
sharing uart. This could re reworked to follow the logic that they
have so that we can upstream the change.

For more details look here:
https://review.trustedfirmware.org/c/TF-M/trusted-firmware-m/+/33529

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>
Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>