[nrf toup] secure_fw: Add option to log output on a shared UART instance.
Add an option to send the log output from the secure firmware on a
UART instance that would be shared with the non-secure application.

This option is added where the number of UART instances is limited
and the application only cares about the receiving the TF-M log
on fatal errors.

To allow this option to be enabled the log is disabled in the boot
process before the non-secure application is started.
It is enabled again when an unrecoverable exception has occurred in
the secure firmware.

NCSDK-18595

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit 19403a8d5dd7fe4bde0e959cf1d242ea4e6f5740)
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit 54af7a244b1aab7dbfdf9fc2cf25440248639e3e)
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>
Change-Id: I65e33f48bd7c6334d04b528c28e8b2d4a3331d0d
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>