[nrf noup] platform: crypto: Check if for config instead of setting it
Check if the required configuration has been set in the configuration
header instead of suddenly setting it outside of the configuration
header.
If the configuration is missing this definition this could lead to an
inconsistent configuration.

This is a noup commit because this definition has been removed in
another upstream commit 7e804909164f523cd995df295c2c4d5a73cb366a and
this is not easily backported.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit 4be04d60d7cd19d7a5ce28352ad6316ff43931ed)
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>