platform: nordic_nrf: Block execution from secondary_partition
Don't allow execution from the secondary partition. There is a concern
that the secondary partition could be written to with malicious code
and then executed from without the image being validated.

It is not clear if it is supported to boot the secondary partition. If
it is then blocking execution like this can not be done.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>