[nrf noup] platform: nordic_nrf: Use aliased default IRQ handlers
Currently we are generating dozens of default IRQ Handlers that each
just spin. In sum, 256 bytes of IRQ handlers that do nothing.

In this patch we save these bytes by using the alias attribute.

Now there is only one default IRQ handler placed in flash and all the
default IRQ handlers alias to it.

I tried to place the duplicated default_tfm_IRQHandler in the common
startup.c, but it did not work. I am not certain, but I think this is
due to a problem with the way we link in TF-M.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
(cherry picked from commit 3dcb0c3e59738bb00a5061a5538cd72f18b632d7)
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>