[nrf noup] platform: nordic_nrf: Use the correct SPU_IRQHandler
There is a build system bug where the wrong IRQ handler is used for
the SPU.

trusted-firmware-m/platform/ext/target/nordic_nrf/common/core/faults.c
defines the symbol SPU_IRQHandler.

And the intention is for this to be used.

But the linker never checks faults.c for undefined symbols because it
already has a weak symbol for SPU_IRQHandler from the startup file to
satisfy itself.

For the weak symbol to work properly faults.c needs to be linked into
the build either because it provides some symbol that is undefined, or
through the use of --whole-archive as Zephyr does.

Until TF-M uses --whole-archive, weak symbols will not work as
intended.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>