[nrf noup] platform: nrf_nordic: Add temp fix for for secure FPU on 9161
The MDK for nRF9120 used in the nRF9161 target doesn't define the Secure FPU
as it doesn't exist, but for other platforms like the 9160 it has a dummy
define, with an UNUSED field in the type.
The long plan is to get this fixed in the MDK but until then, to make
the nrfxlib 3.1.0 update possible this tempfix is applied.

 Ref: NCSDK-23046

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>
Change-Id: I44042ee9aada99c59a5930440306bb6c40ae4880