[nrf fromlist] platform: Store callee saved register in exception information
Store the callee saved registers in the exception information logging.
We store the current exception frame, which has the registers of the
caller saved registers when the exception occurs, but the callee saved
register information is lost during the exception handling.
This provides us with an incomplete picture of the state at the time
the exception occurred.

Upstream PR:
https://review.trustedfirmware.org/c/TF-M/trusted-firmware-m/+/24877

Change-Id: I3d15f9eccf1aa8c2c1b99e75e38229ab82420f36
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>