[nrf fromtree] platform: Store callee saved register in exception information
Store the callee saved registers in the exception information logging.
We store the current exception frame, which has the registers of the
caller saved registers when the exception occurs, but the callee saved
register information is lost during the exception handling.
This provides us with an incomplete picture of the state at the time
the exception occurred.

Change-Id: I3d15f9eccf1aa8c2c1b99e75e38229ab82420f36
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit dbdcfa06699e31ac29f0cc36b54c1f676e4a772e)
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>