[nrf fromlist] platform: nordic_nrf: Remove Flash_Uninitialize function
[nrf fromlist] because this was cherry-picked from
https://review.trustedfirmware.org/c/TF-M/trusted-firmware-m/+/25903/1

Remove the function ARM_Flash_Uninitialize is it is only used in BL1
and BL2 integration and nordic is not compatible with BL1 or BL2.

It is bad practice to have an unused function available as it is
untested and may therefore be unreliable if TF-M were to start to
using it.

It also bloats the size of the binary and has a code maintenance cost.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
Change-Id: I319b13b916e78f7692fab23a4f542877b8484bcb