[nrf noup] Remove check for P256M
TF-M checks if P256M is available during build time using
MBEDCRYPTO_PATH which is set to the TF-M repo to use custom
mbed TLS cmake configurations, but this means the script can not be
found. But as mbed TLS software crypto is not used anyway we can
hardcode P256M to be disabled.

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>
Change-Id: I94fde1f41e3493e840823cae284256176a364863
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>