[nrf fromlist] platform: nordic_nrf: Support for building nRF54L15 + sync repos
This is a big change with two goals:
1) Add initial support for building nRF54L15 with upstream TF-M
2) Align the Nordic platform code between the upstream TF-M and
   Nordics TF-M fork

This change does NOT add support for nRF54L15 in upstream TF-M yet,
it only adds building support at the moment. More effort is needed
to allow running upstream TF-M in this platform. Most of the
configuration files for nRF54L15 are plain copies from the nRF5340
with light modifications to allow building.

This change brings an updated version of the nrfx library as well,
since it is needed to provide definitions for the nRF54L15.

Upstream PR: https://review.trustedfirmware.org/c/TF-M/trusted-firmware-m/+/33956

Change-Id: I7543296f2ba839c5dd886fbc1231a5fedc23fd8f
Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>