[nrf noup]platform: ext: nordic: Add MPC address filter
fixup! [nrf noup] platform: nordic_nrf: Add support for 54l

MPC triggers a MEMACCERR even when address 0xFFFFFFFE is
accessed. This is an issue since this addres is very commonly
acccesed by GDB because of some return values in TF-M.

This update the MPC handler to return execution if it
is triggered for this address.

Ref: NCSDK-28128

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>