[nrf noup] platform: nordic_nrf: Remove unused code
fixup! [nrf noup] platform: nordic_nrf: Add support for 54l

Removed an ifdef which was never used because it was included
in the else statement with the same symbol.

In nRF54L by default we configure all peripherals as non-secure and
then we excplicitely configure as secure the ones needed. So this
was in other case not needed.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>