[nrf fromtree] Platform: corstone1000: Fix isolation L2 memory protection
The whole of the SRAM was configured unprivileged on this platform, so
the memory protection required for isolation level 2 was not present.

This patch changes the S_DATA_START to S_DATA_LIMIT MPU region to be
configured for privileged access only. It also reorders the MPU regions
so that the App RoT sub-region overlapping S_DATA has a higher region
number and so takes priority in the operation of the Armv6-M MPU.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: Icdf169f92f7a47b27ea38dac4098e3205af7f5af
(cherry picked from commit 66596b4dae57bd675d006f4ad3fa3e5db512eba8)
Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>