[nrf noup] platform: nordic_nrf: 54l15: clean up RRAM write buffer size setting
CONFIG_NRF_RRAM_WRITE_BUFFER_SIZE is not passed to TF-M, so it cannot
be used.
Moreover, it seems that it's not even defined on the NS image when
using TF-M.

As a bonus, do not run the COMMIT_WRITEBUF task when using unbuffered writing.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>
(cherry picked from commit 910bdc90593c638633374e8ada71d34909e85d5a)