[nrf noup] ITS: fix aad_size parameter spelling
This is a noup because the upstream commit
77aabb76e4bb1cfa13d1caf913e8335a489d3540
(from the main branch) does not apply cleanly
and would require bringing in too many changes.

Fix the use of the aad_size struct parameter in Nordic's ITS encryption
HAL.

In addition, fix the remaining use of add/add_size as parameter names
in the ITS partition code.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>