[nrf fromtree] platform: nordic_nrf: Fix broken path for NS in nRF54L
Fix a path issue with the non secure application config.cmake
file. The config.cmake for nRF54L10 and nRF54L15 is the same.
To avoid duplicating the file the config.cmake is placed in a
common nrf54l folder and it is included by each platforms
config.cmake files. This doesn't work for the non secure application
because the path does not exist. To fix that install to the non secure
application the common file directly.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: I36d6bed5d1f2a71ed12f9004485accfcbca1b0d6
(cherry picked from commit 51352be0fde5b1d318bbe26ab03c08227542f399)