[nrf noup] inteface: Remove headers existing in the Oberon PSA core
There are multiple headers which exist in the Oberon PSA core and
in TF-M. At the same time some of these headers include other headers
with quotes "" which means that the folder include order doesn't have
any effect.

Instead of relying to the include order of the folders remove the
duplicate files from TF-M since these are not/should not be used.

I removed them with a bash command, just in case is needed:
for i in $(PATH_TO_OBERON/include/psa -name "*.h" -printf "%f\n");
do
    rm TFM_PATH/interface/include/psa/$i;
done

This commit removes headers which are updated by the commit:
[nrf noup] crypto: Add PAKE support

The commits cannot be combined because the have different scope
but in the next upmerge this commit can be applied before the
PAKE support and this will reduce the scope of the PAKE noup.

Ref: NCSDK-33148

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>