[nrf noup] build: fix TF-M version
The TF-M build system retrieves the latest tag to figure out the TF-M
version.
It ends up being wrong because in some cases we don't have the latest
tags from upstream
For example for TF-M 2.1.2 the latest tag was TF-Mv2.1.0, probably
because between the two versions upmerges have been done with
cherry picks instead of upstream tag merging.

Completely stop relying on the tags and only use TFM_VERSION_MANUAL as
the version.
This fixes the TF-M version printed on boot.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>