[nrf noup] build: fix TF-M version
The TF-M build system retrieves the latest tag to figure out the TF-M
version.
It ends up being wrong because for example as of TF-M 2.1.2 in the commit
history the latest tag is TF-Mv2.1.0.
We don't have tags from upstream, but for some reason still have some of
them, which makes it confusing.

Completely stop relying on the tags and only use TFM_VERSION_MANUAL as
the version.
This fixes the TF-M version printed on boot.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>