[nrf noup] Cherry pick provisioning update
This is a partial cherry pick of the commit with the following
details in upstream TF-M:

Platform: Improve tfm_plat_provisioning_is_required signature
Change-Id: I5a35dd477d3f017401d147cf99d07f1537373ac9

The commit brings many changes but here I only picked
the changes that affect the target_cfg.c file of the
Nordic platforms.

This is done as preparation work so that I can sync the
target_cfg.c file in the upstream repo and here and thus
I can cleanly cherry pick subsequent commits from the
upstream repo cleanly.

The current noup commit will be dropped in the next
upmerge so this is a short-lived noup.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>