[nrf noup] secure_fw: partitions: Remove its dependency
TFM_CRYPTO depends on TFM_INTERNAL_TRUSTED_STORAGE_SERVICE.
This means it is not possible to not use ITS.
This is removed to make it possible to support using crypto
without ITS.
This is a noup as it is not possible to do this change upstream.
There are platforms upstream that depend on this dependency.

Signed-off-by: Dag Erik Gjørvad <dag.erik.gjorvad@nordicsemi.no>