[nrf noup] Revert "build: Remove forced NS init and startup files"
This reverts commit 5206a413eb1e51a9ede1701d4102be6446b9299d.

This change does not work with NRF devices. The issue is that
this changes the target which contains the startup files.
Before it was the tfm_ns and this made it to platform_ns.
This does't work with NRF devices, after this change
the startup files are compiled multiple times and for
some unknown reason one of the compilication doesn't work
and produces an .o file which is not an object file.

This needs further investigation, so I opened a task
to investigate further: NCSDK-35302

I reverted the commit to unblock the current upmerge
process.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>