[tobe fromlist] platform: nordic: Update the nrf-pinctrl.h
Update the header nrf-pinctrl.h header to match the latest
one from Zephyr. This file was copied and modified before
in order to inlcude two relevant helper macros.

I created another header called tfm-pinctrl.h which includes
these two macros and the header nrf-pinctl.h. This makes it
easier to understand that this is a custom header and also
allows to copy the nrf-pinctrl.h as is.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>