[nrf fromlist] platform: nordic_nrf: Fix nrf54l initialization issue
The function nordicsemi_nrf54l_init was intented to be used
only for builds outside of sdk-nrf. This was not the case,
the weak attribute used here doesn't seem to have any effect
and this function was used for all builds.

To overcome this make sure that this function is not included
for sdk-nrf builds.

Upstream PR:
https://review.trustedfirmware.org/c/TF-M/trusted-firmware-m/+/45054

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: I0f7a92db95363cf8c0b1d3e9f198675de0705d3b