[nrf noup] platform: nordic_nrf: Fix bug in nrfx errata
nrfx have some functions which dynamically checks if an errata
should be applied based on a value in FICR. It seems that
reading two values from FICR consecutively has undefined
behavior if each call is not followed by a __DSB call.

This commit fixes the issue by copying the nrfx header with the
errata functions and adding a __DSB call to each FICR read.

This is a (hopefully) short lived noup since the fix will come
in nrfx.

Ref: NCSDK-36641

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>