[nrf noup] tf-m: Refactor support for external PSA core
nrf-quash! [nrf noup] tf-m: Enable support for external PSA core

The cmake files which handle the external PSA core are in the
nrf_security folder of sdk-nrf. Instead of using a relative path
to include these files into the TF-M build use the new variable
which points to the tfm_boards folder in sdk-nrf.

Ref: NCSDK-36186

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>